refactor(myProfile): 내 프로필 화면 관련 코드 리팩토링
