+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/myProfile/MyProfileViewModel.kt
@@ -53,21 +53,17 @@ class MyProfileViewModel(
     }
 
     fun onErrorMessageViewed() {
-        _uiState.postValue(
-            uiState.value!!.copy(
-                isError = false,
-                errorMessage = "",
-            ),
+        _uiState.value = uiState.value!!.copy(
+            isError = false,
+            errorMessage = "",
         )
     }
 
     private fun changeErrorUiState(errorMessage: String) {
-        _uiState.postValue(
-            uiState.value!!.copy(
-                isLoading = false,
-                isError = true,
-                errorMessage = errorMessage,
-            ),
+        _uiState.value = uiState.value!!.copy(
+            isLoading = false,
+            isError = true,
+            errorMessage = errorMessage,
         )
     }
